adding Player...

+++ MuffinProject/Assets/Scripts/Refactor/GameEvents.cs
@@ -5,7 +5,7 @@ public static class GameEvents
     public static event Action OnTurnChanged;
     public static event Action<int, int> OnDrawn; // ActorNumber, CardID
     public static event Action<bool> OnHandModeChanged;
-    public static event Action<int> OnHPChanged;
+    public static event Action<float> OnHPChanged;
     public static event Action<int> OnCurrentHandCountChanged;
 
     public static void RaiseTurnChanged()
@@ -23,7 +23,7 @@ public static class GameEvents
         OnHandModeChanged?.Invoke(st);
     }
 
-    public static void RaiseHpChanged(int CurrentHP)
+    public static void RaiseHpChanged(float CurrentHP)
     {
         OnHPChanged?.Invoke(CurrentHP);
     }
